api/requirements.txt
tensorflow==2.10.0
tensorflow-serving-api==2.10.0
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy